feat: add gcloudRun and use getExecOutput (#77)
This introduces two new functions `gcloudRun` and `gcloudRunJSON` which use `@actions/exec#getExecOutput` to execute gcloud commands. It checks if the command was successful. On failure, it throws an error and prints stderr and stdout. On success, it returns the stderr, stdout, and combined output.

Introducing this function also greatly simplified some of the other functions, which had a lot of duplicate logic, etc.