Make getLatestGcloudSDKVersion stubbable and expose TestToolCache (#134)
Exporting es6 modules from within sub-files doesn't work with sinon's
spies, which means all our tests still try to contact the Internet when
running getLatestGcloudSDKVersion, even if the method is stubbed. I
figured this out while traveling on an airplane without an Internet
connection.

This moves the function into the main module so it can be stubbed
appropriately.

This also refactors and exports the `TestToolCache` so we can use it
elsewhere.